requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
distlib==0.3.7
distro==1.8.0
filelock==3.13.1
Flask==2.3.2
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.106.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
gspread==5.11.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.1.1
packaging==23.2
pandas==2.0.3
platformdirs==4.0.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
PyDrive==1.3.1
pydub==0.25.1
pymssql==2.2.8
pyodbc==4.0.39
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
tqdm==4.65.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.7
Werkzeug==2.3.6
yarl==1.9.2